multipart: use same host as main module
The multipart component of the SDK doesn't use the same request
mechanism as the rest of the library. While the rest of the library
already had a way to change the hostname/port of the API server, the
multipart code had hardcoded it.

This adds a separate method to the Kloudless object for getting the API
endpoint, and then uses it for multipart.

An example of using multipart with a non-standard host:

    var kl = require('kloudless')('your-api-key'),
        fs = require('fs');

    kl.setHost('example.com', 443); // only this line changes
    kl.files.uploadMultipart({
      name: '05-Televisor-Pinup.mp3',
      account_id: 'your-account-id',
      parent_id: 'your-parent-id',
      file: fs.createReadStream('05-Televisor-Pinup.mp3')
    }, function() { console.log('upload complete'); });

N.B. it's possible to specify a protocol (presumably http), but it's
completely ignored by the code, which does things like
    var req = https.request({ ...
So you can really only use https right now.